record

+++ Visualization/VirtualMemory/Utility.hpp
@@ -11,8 +11,9 @@ struct Vector3 {
 		X(x), Y(y), Z(z) {}
 
 	template<typename T = int>
-	Vector3(T x = 0, T y = 0, T z = 0) :
+	Vector3(T x = (int)0, T y = (int)0, T z = (int)0) :
 		X(x), Y(y), Z(z) {}
+
 
 	bool operator == (const Vector3 &vector)const {
 		return (X == vector.X) && (Y == vector.Y) && (Z == vector.Z);
